Remove state from cli component
